refactor(Makefile): Tidy up and remove redundancy
Refactored Makfile to tidy up the recipes and remove repetition

+++ Makefile
@@ -3,7 +3,7 @@
 #
 .EXPORT_ALL_VARIABLES:
 
-.PHONY: yaml help clean clean-configmaps clean-deployments clean-services clean-ingress apply-configmaps apply-deployments apply-services apply-ingress enable-ingress
+.PHONY: yaml help clean $(K_CONFIGS) $(YAML_TARGETS) clean-configmaps clean-deployments clean-services clean-ingress apply-configmaps apply-deployments apply-services apply-ingress enable-ingress
 
 #set default ENV based on your username and hostname
 TOUCH_FILES=enable-ingress
@@ -12,6 +12,7 @@ TEMP_FOLDERS=
 K_ROOT := ops/kubernetes
 YAML_FILES_SRC = $(shell find $(K_ROOT) -type f -name '*.ytemplate')
 YAML_FILES = $(YAML_FILES_SRC:%.ytemplate=%.yaml)
+YAML_TARGETS = $(YAML_FILES_SRC:%.ytemplate=%)
 
 # We do these separately so we can order them.
 K_CONFIGS := $(wildcard $(K_ROOT)/configmaps/*.yaml $(K_ROOT)/deployments/*.yaml $(K_ROOT)/services/*.yaml $(K_ROOT)/ingress/*.yaml)
@@ -37,23 +38,11 @@ endef
 help:
 	@grep -E '^[0-9a-zA-Z_-]+:.*?## .*$$' $(MAKEFILE_LIST) | sort | awk 'BEGIN {FS = ":.*?## "}; {printf "\033[36mmake %-30s\033[0m %s\n", $$1, $$2}'
 
-apply-all: enable-ingress apply-configmaps apply-deployments apply-services apply-ingress ## Apply all resources to kubernetes
+apply-all: enable-ingress apply ## Apply all resources to kubernetes
 
-apply-configmaps: $(K_CONFIGMAPS)/*.yaml ## apply the configmaps to kubernetes
-	@echo "Applying configmap in $^"
-	$(APPLYYML)
+apply: $(YAML_TARGETS)
 
-apply-deployments: $(K_DEPLOYMENTS)/*.yaml ## apply the deployments to kubernetes
-	@echo "Applying deployments in $^"
-	$(APPLYYML)
-
-apply-services: $(K_SERVICES)/*.yaml ## apply the services to kubernetes
-	@echo "Applying services in $^"
-	$(APPLYYML)
-
-apply-ingress: $(K_INGRESS)/*.yaml ## apply the ingress to kubernetes
-	@echo "Applying ingress in $^"
-	$(APPLYYML)
+apply-configmaps: $(K_CONFIGS)  ## apply the configmaps to kubernetes
 
 enable-ingress: ## enable ingress addon in minikube
 	minikube addons enable ingress
@@ -76,8 +65,19 @@ enable-ingress: ## enable ingress addon in minikube
 # 	$(DELETE_RESOURCE)	
 
 clean:  ## clean everything up
-	-rm $(TOUCH_FILES)
-	-rm -Rf $(TEMP_FOLDERS)
+	-rm $(TOUCH_FILES) || true
+	@[ "${var}" ] && rm -Rf $(TEMP_FOLDERS) || true
+
+	## iterate through the yaml files and delete the resources before we
+	## clean the actual yaml files away forever
+	( \
+		array=($$(echo "$$YAML_FILES" | tr ' ' '\n')); \
+		for item in "$${array[@]}"; do \
+			echo "deleting resources in ''$$item'"; \
+			kubectl delete -f $$item || true; \
+		done \
+	)
+
 	-rm $(YAML_FILES)
 
 $(APPS): ## restart the apps
@@ -93,7 +93,9 @@ lint-requirements: ## install kubernetes linter
 
 lint: lint-requirements ## run lint checker against the yaml files
 	@echo "Running Linter"
-	kube-linter lint $(K_ROOT)
+	kube-linter lint $(K_ROOT) || true # just so make doesn't blob when there is nothing we can do *shrug*
+	# as in case of the "no pods found matching service labels (map[app:helloworld])" which is only true because
+	# the linter is not looking at different files... 
 
 yaml: $(YAML_FILES) ## generate yaml files from template (subst env variables)
 
@@ -101,3 +103,7 @@ yaml: $(YAML_FILES) ## generate yaml files from template (subst env variables)
 	@echo "Creating yaml files"
 	@echo "Creating file $@ from $<"
 	export K_NAMESPACE=$(K_NAMESPACE) && envsubst < $< > $@
+
+%: %.yaml
+	@echo "Applying yaml file $<"
+	kubectl apply -f $^
